third_party/blink/web_tests/accessibility/aria-flowto.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body id="body">

<h1 id="item1" aria-flowto="item2 item3"></h1>

<div id="item2">Content 2</div>
<div id="item3">Content 3</div>

<p id="description"></p>
<div id="console"></div>

<script>
test((t) => {
    var item1 = accessibilityController.accessibleElementById("item1");
    var item2 = accessibilityController.accessibleElementById("item2");
    var item3 = accessibilityController.accessibleElementById("item3");

    assert_true(item1.ariaFlowToElementAtIndex(0).isEqual(item2));
    assert_true(item1.ariaFlowToElementAtIndex(1).isEqual(item3));
},"This tests that aria-flowto identifies the correct element");
</script>

</body>
</html>
